requirements.txt
Django==2.1.1
django-extensions
gunicorn
ipdb
psycopg2
psycopg2-binary
whitenoise
djangorestframework
drf-yasg[validation]==1.9.1
drf-writable-nested
social-auth-core
social-auth-app-django
flower
gevent
gevent-eventemitter
greenlet
celery
social-auth-core
social-auth-app-django
django-oauth-toolkit==1.0.0
drf-extensions==0.3.1
pillow
tqdm
termcolor
faker